backend/financial_crime_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
elasticsearch
psycopg2-binary
neo4j-driver
python-dotenv